requirements.txt
# Core dependencies
google-generativeai==0.8.3
gradio==3.50.2
Pillow==10.1.0
pdf2image==1.16.3
pydantic==1.10.13
python-poppler==0.4.1

# Additional dependencies that might be required
#numpy==1.26.2
#requests==2.31.0
#urllib3==2.1.0
#protobuf==4.25.1
#packaging==23.2